+++ src/lib_dal_node/cli.ml
@@ -25,6 +25,8 @@
 
 module Types = Tezos_dal_node_services.Types
 
+let exits = Errors.Exit_codes.all
+
 let env_value_starts_with_yes ~env_var =
   match Sys.getenv_opt env_var with
   | None -> false
@@ -707,10 +709,10 @@ let wrap_with_error main_promise =
   let open Lwt_syntax in
   let* r = Lwt_exit.wrap_and_exit main_promise in
   match r with
-  | Ok () -> Lwt_exit.exit_and_wait 0
+  | Ok () -> Lwt_exit.exit_and_wait Errors.Exit_codes.ok
   | Error err ->
       let* () = Lwt_io.eprint (Format.asprintf "%a" pp_print_trace err) in
-      Lwt_exit.exit_and_wait 1
+      Lwt_exit.exit_and_wait Errors.Exit_codes.some_error
 
 (** [wrap_action action] is the main entry point wrapper for DAL node commands.
     It sets up the exception filter to handle all exceptions except runtime ones,
@@ -731,7 +733,7 @@ module Run = struct
 
   let info =
     let version = Tezos_version_value.Bin_version.octez_version_string in
-    Cmdliner.Cmd.info ~doc:"Run the Octez DAL node" ~man ~version "run"
+    Cmdliner.Cmd.info ~exits ~doc:"Run the Octez DAL node" ~man ~version "run"
 
   let action =
    fun data_dir
@@ -947,6 +949,10 @@ module Config = struct
        $ verbose $ ignore_l1_config_peers $ disable_amplification
        $ batching_configuration))
 
+  let print_trace ~error_trace =
+    Format.eprintf "Failure: %a" Error_monad.pp_print_trace error_trace ;
+    Lwt.return_unit
+
   module Init = struct
     let man =
       [
@@ -955,20 +961,53 @@ module Config = struct
           "This command creates a configuration file with the parameters \
            provided on the command-line, if no configuration file exists \
            already in the specified or default location. Otherwise, the \
-           command-line parameters override the existing ones, and old \
-           parameters are lost. This configuration is then used by the run \
-           command.";
+           command fails, use reset instead. This configuration is then used \
+           by the run command.";
       ]
 
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
-      Cmdliner.Cmd.info ~doc:"Configuration initialisation" ~man ~version "init"
+      Cmdliner.Cmd.info
+        ~exits
+        ~doc:"Configuration initialisation"
+        ~man
+        ~version
+        "init"
 
     let action =
       mk_action @@ fun ~config_file ~configuration_override ->
-      Configuration_file.save
-        ~config_file
-        (configuration_override Configuration_file.default)
+      Configuration_file.exit_on_configuration_error ~emit:print_trace
+      @@ Configuration_file.save
+           ~allow_overwrite:false
+           ~config_file
+           (configuration_override Configuration_file.default)
+
+    let term = mk_term action
+
+    let cmd = Cmdliner.Cmd.v info term
+  end
+
+  module Reset = struct
+    let man =
+      [
+        `S "DESCRIPTION";
+        `P
+          "This command replace the existing configuration file in the \
+           specified or default location with the parameters provided on the \
+           command-line. This configuration is then used by the run command.";
+      ]
+
+    let info =
+      let version = Tezos_version_value.Bin_version.octez_version_string in
+      Cmdliner.Cmd.info ~exits ~doc:"Configuration reset" ~man ~version "reset"
+
+    let action =
+      mk_action @@ fun ~config_file ~configuration_override ->
+      Configuration_file.exit_on_configuration_error ~emit:print_trace
+      @@ Configuration_file.save
+           ~allow_overwrite:true
+           ~config_file
+           (configuration_override Configuration_file.default)
 
     let term = mk_term action
 
@@ -987,13 +1026,22 @@ module Config = struct
 
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
-      Cmdliner.Cmd.info ~doc:"Configuration update" ~man ~version "update"
+      Cmdliner.Cmd.info
+        ~exits
+        ~doc:"Configuration update"
+        ~man
+        ~version
+        "update"
 
     let action =
       mk_action @@ fun ~config_file ~configuration_override ->
       let open Lwt_result_syntax in
-      let* configuration = Configuration_file.load ~config_file in
+      let* configuration =
+        Configuration_file.exit_on_configuration_error ~emit:print_trace
+        @@ Configuration_file.load ~config_file ()
+      in
       Configuration_file.save
+        ~allow_overwrite:true
         ~config_file
         (configuration_override configuration)
 
@@ -1007,12 +1055,13 @@ module Config = struct
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
       Cmdliner.Cmd.info
+        ~exits
         ~doc:"Manage the Octez DAL node configuration"
         ~man
         ~version
         "config"
     in
-    Cmdliner.Cmd.group ~default info [Init.cmd; Update.cmd]
+    Cmdliner.Cmd.group ~default info [Init.cmd; Reset.cmd; Update.cmd]
 end
 
 module Snapshot = struct
@@ -1060,7 +1109,7 @@ module Snapshot = struct
 
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
-      Cmdliner.Cmd.info ~doc:"Export snapshot" ~man ~version "export"
+      Cmdliner.Cmd.info ~exits ~doc:"Export snapshot" ~man ~version "export"
 
     let action min_published_level max_published_level slots data_dir endpoint
         config_file file_path =
@@ -1107,7 +1156,7 @@ module Snapshot = struct
 
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
-      Cmdliner.Cmd.info ~doc:"Import snapshot" ~man ~version "import"
+      Cmdliner.Cmd.info ~exits ~doc:"Import snapshot" ~man ~version "import"
 
     let action min_published_level max_published_level slots data_dir endpoint
         config_file no_check file_path =
@@ -1148,7 +1197,12 @@ module Snapshot = struct
     let default = Cmdliner.Term.(ret (const (`Help (`Pager, None)))) in
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
-      Cmdliner.Cmd.info ~doc:"Snapshot management" ~man ~version "snapshot"
+      Cmdliner.Cmd.info
+        ~exits
+        ~doc:"Snapshot management"
+        ~man
+        ~version
+        "snapshot"
     in
     Cmdliner.Cmd.group ~default info [Export.cmd; Import.cmd]
 end
@@ -1179,7 +1233,12 @@ module Debug = struct
 
         let info =
           let version = Tezos_version_value.Bin_version.octez_version_string in
-          Cmdliner.Cmd.info ~doc:"Print SQL statements" ~man ~version "schemas"
+          Cmdliner.Cmd.info
+            ~exits
+            ~doc:"Print SQL statements"
+            ~man
+            ~version
+            "schemas"
 
         let action () =
           Lwt_utils_unix.with_tempdir "store" @@ fun data_dir ->
@@ -1199,6 +1258,7 @@ module Debug = struct
         let info =
           let version = Tezos_version_value.Bin_version.octez_version_string in
           Cmdliner.Cmd.info
+            ~exits
             ~doc:"Print DAL node store debug information"
             ~man
             ~version
@@ -1211,7 +1271,12 @@ module Debug = struct
       let default = Cmdliner.Term.(ret (const (`Help (`Pager, None)))) in
       let info =
         let version = Tezos_version_value.Bin_version.octez_version_string in
-        Cmdliner.Cmd.info ~doc:"Print debug information" ~man ~version "print"
+        Cmdliner.Cmd.info
+          ~exits
+          ~doc:"Print debug information"
+          ~man
+          ~version
+          "print"
       in
       Cmdliner.Cmd.group ~default info [Store.cmd]
   end
@@ -1220,7 +1285,7 @@ module Debug = struct
     let default = Cmdliner.Term.(ret (const (`Help (`Pager, None)))) in
     let info =
       let version = Tezos_version_value.Bin_version.octez_version_string in
-      Cmdliner.Cmd.info ~doc:"Debug commands" ~man ~version "debug"
+      Cmdliner.Cmd.info ~exits ~doc:"Debug commands" ~man ~version "debug"
     in
     Cmdliner.Cmd.group ~default info [Print.cmd]
 end
@@ -1326,7 +1391,7 @@ let commands =
   let default = Cmdliner.Term.(ret (const (`Help (`Pager, None)))) in
   let info =
     let version = Tezos_version_value.Bin_version.octez_version_string in
-    Cmdliner.Cmd.info ~doc:"The Octez DAL node" ~version "octez-dal-node"
+    Cmdliner.Cmd.info ~exits ~doc:"The Octez DAL node" ~version "octez-dal-node"
   in
   Cmdliner.Cmd.group
     ~default