+++ src/lib_dal_node/configuration_file.mli
@@ -108,10 +108,28 @@ val store_path : t -> string
     from a data dir. *)
 val default_config_file : string -> string
 
-(** [save config] writes config file in [config.data_dir] *)
-val save : config_file:string -> t -> unit tzresult Lwt.t
+(** [save ~allow_overwrite ~config_file] writes config file in [config.data_dir] if it doesn't exists or allow_overwrite is true.
 
-val load : config_file:string -> t tzresult Lwt.t
+    If the file exists and allow_overwrite is false, the function returns an error.
+*)
+val save :
+  allow_overwrite:bool -> config_file:string -> t -> unit tzresult Lwt.t
+
+(** [load ~on_file_not_found ~config_file ()] load config file from
+    [config.data_dir] if it exists. If not it calls the provided on_file_not_found function.
+    If no such function is provided, an error is returned.
+*)
+val load :
+  ?on_file_not_found:(unit -> t tzresult Lwt.t) ->
+  config_file:string ->
+  unit ->
+  t tzresult Lwt.t
+
+(* Helper to exit with a proper code when a configuration error occurs *)
+val exit_on_configuration_error :
+  emit:(error_trace:tztrace -> experimental_features Lwt.t) ->
+  'a tzresult Lwt.t ->
+  'a tzresult Lwt.t
 
 (** [identity_file config] returns the absolute path to the
     "identity.json" file of the DAL node, based on the configuration