+++ src/lib_dal_node/event.ml
@@ -91,6 +91,26 @@ open struct
       ~level:Notice
       ("exit_status", Data_encoding.int8)
 
+  let config_load_fail =
+    declare_1
+      ~section
+      ~prefix_name_with_section:true
+      ~name:"config_load_fail"
+      ~msg:"Loading configuration failed with error {error} "
+      ~level:Error
+      ~pp1:Error_monad.pp_print_trace
+      ("error", Error_monad.trace_encoding)
+
+  let config_save_fail =
+    declare_1
+      ~section
+      ~prefix_name_with_section:true
+      ~name:"config_save_fail"
+      ~msg:"saving configuration failed with error {error} "
+      ~level:Error
+      ~pp1:Error_monad.pp_print_trace
+      ("error", Error_monad.trace_encoding)
+
   let dal_node_sqlite3_store_init =
     declare_0
       ~section
@@ -1413,6 +1433,12 @@ let emit_history_mode_warning ~stored_levels ~storage_period =
   emit history_mode_warning (stored_levels, storage_period)
 
 let emit_configuration_loaded () = emit configuration_loaded ()
+
+let emit_configuration_loading_failed ~error_trace =
+  emit config_load_fail error_trace
+
+let emit_configuration_saving_failed ~error_trace =
+  emit config_save_fail error_trace
 
 let emit_upgrading_configuration ~from ~into =
   emit upgrading_configuration (from, into)