+++ tezt/lib_tezos/dal_node.mli
@@ -174,6 +174,20 @@ val check_error : ?exit_code:int -> ?msg:Base.rex -> t -> unit Lwt.t
     running, make the test fail. *)
 val wait : t -> Unix.process_status Lwt.t
 
+val spawn_config_init :
+  ?expected_pow:float ->
+  ?peers:string list ->
+  ?attester_profiles:string list ->
+  ?operator_profiles:int list ->
+  ?observer_profiles:int list ->
+  ?bootstrap_profile:bool ->
+  ?history_mode:history_mode ->
+  ?slots_backup_uris:string list ->
+  ?trust_slots_backup_uris:bool ->
+  ?batching_time_interval:string ->
+  t ->
+  Process.t
+
 (** Run [octez-dal-node config init].
 
     [expected_pow] allows to change the PoW difficulty. Default value is 0.
@@ -192,6 +206,52 @@ val init_config :
   t ->
   unit Lwt.t
 
+val spawn_config_reset :
+  ?expected_pow:float ->
+  ?peers:string list ->
+  ?attester_profiles:string list ->
+  ?operator_profiles:int list ->
+  ?observer_profiles:int list ->
+  ?bootstrap_profile:bool ->
+  ?history_mode:history_mode ->
+  ?slots_backup_uris:string list ->
+  ?trust_slots_backup_uris:bool ->
+  ?batching_time_interval:string ->
+  t ->
+  Process.t
+
+(** Run [octez-dal-node config init].
+
+    [expected_pow] allows to change the PoW difficulty. Default value is 0.
+*)
+val reset_config :
+  ?expected_pow:float ->
+  ?peers:string list ->
+  ?attester_profiles:string list ->
+  ?operator_profiles:int list ->
+  ?observer_profiles:int list ->
+  ?bootstrap_profile:bool ->
+  ?history_mode:history_mode ->
+  ?slots_backup_uris:string list ->
+  ?trust_slots_backup_uris:bool ->
+  ?batching_time_interval:string ->
+  t ->
+  unit Lwt.t
+
+val spawn_config_update :
+  ?expected_pow:float ->
+  ?peers:string list ->
+  ?attester_profiles:string list ->
+  ?operator_profiles:int list ->
+  ?observer_profiles:int list ->
+  ?bootstrap_profile:bool ->
+  ?history_mode:history_mode ->
+  ?slots_backup_uris:string list ->
+  ?trust_slots_backup_uris:bool ->
+  ?batching_time_interval:string ->
+  t ->
+  Process.t
+
 val update_config :
   ?expected_pow:float ->
   ?peers:string list ->
@@ -208,6 +268,9 @@ val update_config :
 
 module Config_file : sig
   (** DAL node configuration files. *)
+
+  (**   *)
+  val filename : t -> string
 
   (** Read the configuration file ([config.json]) of a DAL node. *)
   val read : t -> JSON.t